requirements.txt
# ChaosClaw Dependencies

# Ethereum
web3>=6.0.0
eth-account>=0.8.0

# Twitter/X API
tweepy>=4.14.0

# ChaosChain SDK (for reputation lookups)
chaoschain-sdk>=0.4.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.0.0

# Async
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
